+++ js/dataframe/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @gradio/dataframe
 
+## 0.8.7
+
+### Dependency updates
+
+- @gradio/upload@0.10.7
+- @gradio/client@0.20.1
+- @gradio/button@0.2.40
+
+## 0.8.6
+
+### Dependency updates
+
+- @gradio/client@0.20.0
+- @gradio/statustracker@0.6.0
+- @gradio/button@0.2.39
+- @gradio/upload@0.10.6
+- @gradio/markdown@0.7.6
+
 ## 0.8.5
 
 ### Dependency updates