+++ js/lite/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @gradio/lite
 
+## 4.32.2
+
+### Dependency updates
+
+- gradio@4.32.2
+
+## 4.32.1
+
+### Dependency updates
+
+- gradio@4.32.1
+
+## 4.32.0
+
+### Dependency updates
+
+- gradio@4.32.0
+
 ## 4.31.5
 
 ### Dependency updates